# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deriv Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List trades
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, cancelled, pending]
        - name: type
          in: query
          schema:
            type: string
            enum: [vanilla_options, forex, multiplier]
        - name: symbol
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeList"

    post:
      operationId: createTrade
      tags: [trades]
      summary: Open a new trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Trade opened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/recent:
    get:
      operationId: getRecentTrades
      tags: [trades]
      summary: Get recent trade activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recent trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
  /trades/stats:
    get:
      operationId: getTradeStats
      tags: [trades]
      summary: Trading statistics by type, asset, and period
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, all]
      responses:
        "200":
          description: Trade statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeStats"
  /trades/{id}:
    get:
      operationId: getTrade
      tags: [trades]
      summary: Get a specific trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "404":
          description: Not found

    patch:
      operationId: updateTrade
      tags: [trades]
      summary: Update trade (target profit, notes, strategy)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeUpdate"
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"

    delete:
      operationId: closeTrade
      tags: [trades]
      summary: Close or cancel a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}/logs:
    get:
      operationId: listTradeLogs
      tags: [trades]
      summary: Get logs for a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeLog"

    post:
      operationId: addTradeLog
      tags: [trades]
      summary: Add a log entry to a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeLogInput"
      responses:
        "201":
          description: Log added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeLog"
  /trades/{id}/comments:
    get:
      operationId: listTradeComments
      tags: [trades]
      summary: Get comments on a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trade comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TradeComment"

    post:
      operationId: addTradeComment
      tags: [trades]
      summary: Add a comment to a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeCommentInput"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeComment"
  /trades/{id}/comments/{commentId}:
    delete:
      operationId: deleteTradeComment
      tags: [trades]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Strategies ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        symbol:
          type: string
        displayName:
          type: string
        type:
          type: string
          enum: [vanilla_options, forex, multiplier]
        direction:
          type: string
          enum: [buy, sell, call, put]
        stake:
          type: number
        targetProfit:
          type: ["number", "null"]
        currentProfit:
          type: ["number", "null"]
        entryPrice:
          type: ["number", "null"]
        exitPrice:
          type: ["number", "null"]
        status:
          type: string
          enum: [open, closed, cancelled, pending, closing]
        contractId:
          type: ["string", "null"]
        strategyId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        aiConfirmed:
          type: boolean
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: ["string", "null"]
          format: date-time
        duration:
          type: ["integer", "null"]
        durationUnit:
          type: ["string", "null"]
        mode:
          type: string
          enum: [demo, live]
      required:
        - id
        - userId
        - symbol
        - displayName
        - type
        - direction
        - stake
        - status
        - aiConfirmed
        - openedAt
        - mode
    TradeList:
      type: object
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - trades
        - total
        - page
        - limit
    TradeInput:
      type: object
      properties:
        symbol:
          type: string
        type:
          type: string
          enum: [vanilla_options, forex, multiplier]
        direction:
          type: string
          enum: [buy, sell, call, put]
        stake:
          type: number
        targetProfit:
          type: ["number", "null"]
        strategyId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        durationUnit:
          type: ["string", "null"]
        aiConfirmed:
          type: boolean
        mode:
          type: string
          enum: [demo, live]
        barrier:
          type: ["string", "null"]
          description: >
            Strike/barrier for vanilla options. Use "+0.00" for at-the-money,
            "+N" / "-N" for relative pips from spot, or an absolute price string.
            Required for VANILLALONGCALL / VANILLALONGPUT contract types.
        multiplier:
          type: ["number", "null"]
          description: >
            Multiplier value. Used only for Multiplier trade types. Acceptable values: 40, 100, 200, 300, 400.
      required:
        - symbol
        - type
        - direction
        - stake
        - aiConfirmed
    TradeUpdate:
      type: object
      properties:
        targetProfit:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        strategyId:
          type: ["integer", "null"]
    TradeStats:
      type: object
      properties:
        totalTrades:
          type: integer
        winRate:
          type: number
        totalPnl:
          type: number
        totalStake:
          type: number
        averageProfit:
          type: number
        byType:
          type: array
          items:
            $ref: "#/components/schemas/StatByType"
        byAsset:
          type: array
          items:
            $ref: "#/components/schemas/StatByAsset"
        dailyPnl:
          type: array
          items:
            $ref: "#/components/schemas/DailyPnl"
      required:
        - totalTrades
        - winRate
        - totalPnl
        - totalStake
        - averageProfit
        - byType
        - byAsset
        - dailyPnl
    StatByType:
      type: object
      properties:
        type:
          type: string
        count:
          type: integer
        pnl:
          type: number
        winRate:
          type: number
      required:
        - type
        - count
        - pnl
        - winRate
    StatByAsset:
      type: object
      properties:
        symbol:
          type: string
        displayName:
          type: string
        count:
          type: integer
        pnl:
          type: number
      required:
        - symbol
        - displayName
        - count
        - pnl
    DailyPnl:
      type: object
      properties:
        date:
          type: string
        pnl:
          type: number
        trades:
          type: integer
      required:
        - date
        - pnl
        - trades
    TradeLog:
      type: object
      properties:
        id:
          type: integer
        tradeId:
          type: integer
        level:
          type: string
          enum: [info, warning, error, success]
        message:
          type: string
        metadata:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - tradeId
        - level
        - message
        - createdAt
    TradeLogInput:
      type: object
      properties:
        level:
          type: string
          enum: [info, warning, error, success]
        message:
          type: string
        metadata:
          type: ["string", "null"]
      required:
        - level
        - message
    TradeComment:
      type: object
      properties:
        id:
          type: integer
        tradeId:
          type: integer
        userId:
          type: string
        userDisplayName:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - tradeId
        - userId
        - content
        - createdAt
    TradeCommentInput:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    # ─── Strategies ─────────────────────────────────────────────────────────────
...